+++ core/src/main/java/org/svip/sbomvex/vexstatement/status/VulnStatus.java
@@ -1,0 +1,15 @@
+package org.svip.sbomvex.vexstatement.status;
+
+/**
+ * file: VulnStatus.java
+ * An enumeration that holds all the possible vulnerability
+ * statuses of a VEX statement
+ *
+ * @author Matthew Morrison
+ */
+public enum VulnStatus {
+    NOT_AFFECTED,
+    AFFECTED,
+    FIXED,
+    UNDER_INVESTIGATION
+}